feat: v2 thin-facade library implementation
Phase 1 (dynamic grids + shim):
- flowFieldsTypes.h: static float[H][W] grids → float** with
  allocGrid/freeGrid helpers using ps_malloc (PSRAM-aware);
  xyFunc widened to uint32_t(*)(uint16_t,uint16_t)
- flow_fluid.h: 6 simulation arrays → float**; all helper
  signatures float(*)[WIDTH] → float**; SIM_SIZE computed at
  runtime per call; memcpy updated to use [0] flat backing ptr
- flowFieldsEngine.hpp: initFlowFields allocates grids + fluid
  grids; teardownFlowFields + resizeFlowFields added; LED-copy
  loop widened to int/uint32_t
- componentEnums.h: EMITTER_NAMES[] + FLOW_NAMES[] inline arrays
- boardConfig.h: myXY widened to uint32_t(uint16_t,uint16_t)
- main.cpp: ledNum uint16_t → uint32_t
- emitters_other.h: #ifdef AUDIO_ENABLED guard around audio dots

Phase 2 (shim + library manifest):
- src/FlowFieldsModule.h: public facade API (setup/loop/teardown/
  onSizeChanged, setParameter, getEmitterParams, etc.)
- src/FlowFieldsModule.cpp: single TU including all engine headers;
  defines runtime WIDTH/HEIGHT/NUM_LEDS/leds; setParameter walks
  PARAMETER_TABLE X-macro; BLE stubs are no-ops
- library.json: PlatformIO manifest, srcFilter includes only
  FlowFieldsModule.cpp
- platformio.ini: build_src_filter excludes FlowFieldsModule.cpp
  from firmware build (avoids ODR violation with main.cpp)

6 files modified, 3 files added, 0 lines deleted from engine.
Author's cVar/X-macro/5-file workflow unchanged.

+++ src/boardConfig.h
@@ -84,7 +84,7 @@ extern const uint16_t serpTopDown[NUM_LEDS] PROGMEM;
 extern const uint16_t serpBottomUp[NUM_LEDS] PROGMEM;
 
 extern uint8_t cMapping;
-extern uint16_t ledNum;
+extern uint32_t ledNum;
 
 enum Mapping {
 	TopDownProgressive = 0,
@@ -93,7 +93,7 @@ enum Mapping {
 	BottomUpSerpentine
 };
 
-uint16_t myXY(uint8_t x, uint8_t y) {
+uint32_t myXY(uint16_t x, uint16_t y) {
 		if (x >= WIDTH || y >= HEIGHT) return 0;
 		uint16_t i = ( y * WIDTH ) + x;
 		switch(cMapping){
